refactor: relocate misplaced report files and rename general/ to errors/
