+++ core/lib/report/render/host-table.ts
@@ -1,4 +1,4 @@
-import { markdownTable, type ColumnFormat } from "../../general/markdown-table.ts";
+import { markdownTable, type ColumnFormat } from "./markdown-table.ts";
 import type { AggregatedEntry } from "../../log/aggregate.ts";
 
 export interface HostTableRow extends Omit<AggregatedEntry, "reason"> {
