+++ core/lib/report/render/markdown-table.ts
