+++ setup/docker/transparent/scripts/report-action.node.ts
@@ -13,7 +13,7 @@ import { buildReportParameters } from "../../../../core/lib/report/parameters.ts
 import { buildTransparentReportData } from "../../../../core/lib/report/build/transparent.ts";
 import { renderReportMarkdown } from "../../../../core/lib/report/render/render-report-markdown.ts";
 import { emitBlockedOutcome } from "../../../../core/lib/report/outcome/emit-blocked-outcome.ts";
-import { errorMessage } from "../../../../core/lib/general/error-message.ts";
+import { errorMessage } from "../../../../core/lib/errors/error-message.ts";
 import { writeStepSummary } from "../../lib/write-step-summary.ts";
 
 const LOG_FILE = "/var/log/haproxy/current";
